move main function to main.py
